backend/pyproject.toml
[project]
name = "fastapi-starter"
version = "0.1.0"
description = "FastAPI starter"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard]>=0.115.13",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.9.1",
    "loguru>=0.7.2",
    "sqlmodel>=0.0.27",
    "pyjwt>=2.10.1",
    "alembic>=1.17.2",
    "pymysql[rsa]>=1.1.2",
    "pwdlib[argon2]>=0.3.0",
    "agent-framework-ag-ui>=1.0.0b251223",
]

[project.scripts]
dev = "scripts.commands:dev"
start = "scripts.commands:start"
migrate-dev = "scripts.commands:migrate_dev"
migrate-prod = "scripts.commands:migrate_prod"
make-migrations = "scripts.commands:make_migrations"
lint = "scripts.commands:lint"
test = "scripts.commands:test"

[tool.setuptools.packages.find]
include = ["app*", "scripts*"]
exclude = ["logs*", "alembic*", "openspec*"]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "ruff>=0.14.3",
    "types-pymysql>=1.1.0.20251220",
]
test = [
    "pytest>=8.4.0",
    "pytest-asyncio>=1.2.0",
    "httpx>=0.28.1",
    "pytest-cov>=6.3.0",
]

[tool.uv]
package = true

[tool.mypy]
exclude = ["venv", ".venv", "alembic"]

[tool.ruff]
target-version = "py310"
exclude = ["venv", ".venv", "alembic"]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG001", # unused arguments in functions
    "T201",   # print statements are not allowed
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "W191",  # indentation contains tabs
    "B904",  # Allow raising exceptions without from e, for HTTPException
]

[tool.ruff.lint.pyupgrade]
# Preserve types, even if a file imports `from __future__ import annotations`.
keep-runtime-typing = true

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = [
    "--cov=app",
    "--cov-report=term-missing",
    "--cov-report=html",
]

[tool.coverage.run]
branch = true
source = ["app"]

[tool.coverage.report]
show_missing = true
skip_covered = false
